+++ Gpios.h
@@ -2,13 +2,13 @@
 #define __GPIOS_H
 
 namespace gpios {
-  static constexpr uint8_t kMotorRightEn = 25;
-  static constexpr uint8_t kMotorRightIn1 = 26;
-  static constexpr uint8_t kMotorRightIn2 = 27;
+  static constexpr unsigned int kMotorRightEn = 25;
+  static constexpr unsigned int kMotorRightIn1 = 26;
+  static constexpr unsigned int kMotorRightIn2 = 27;
 
-  static constexpr uint8_t kMotorLeftEn  = 4;
-  static constexpr uint8_t kMotorLeftIn1 = 16;
-  static constexpr uint8_t kMotorLeftIn2 = 17;
+  static constexpr unsigned int kMotorLeftEn  = 4;
+  static constexpr unsigned int kMotorLeftIn1 = 16;
+  static constexpr unsigned int kMotorLeftIn2 = 17;
 }
 
 #endif // __GPIOS_H
